added Message type Channel voice message used by both Midi1 and MIDI 2
Fix char offset for large buffers in profiles and pe

set peRequestIdx properly in processPESysex

+++ src/midi2Profiles.cpp
@@ -102,7 +102,7 @@ void midi2Processor::processProfileSysex(uint8_t group, uint8_t s7Byte){
 
             //******************
 
-            uint8_t charOffset = (syExMessInt[group].pos - 22) % S7_BUFFERLEN;
+            uint16_t charOffset = (syExMessInt[group].pos - 22) % S7_BUFFERLEN;
             uint16_t dataLength = syExMessInt[group].intbuffer1[0];
             if(
                     (syExMessInt[group].pos >= 22 && syExMessInt[group].pos <= 21 + dataLength)
